+++ src/template/ts/src/menu.ts
@@ -75,7 +75,7 @@ export const interpretMenuAction = async (state: AppState): Promise<void> => {
       about: async (state: AppState): Promise<void> => {
         await titleScreen("___APP NAME___");
         console.log(
-          boxen(chalk.yellow(`Author: `) + "YOUR NAME", blankBoxenStyle)
+          boxen(chalk.yellow(`Author: `) + "___AUTHOR NAME___", blankBoxenStyle)
         );
 
         console.log("Press any key to return to Main Menu ...");